update spec, get proper leader
First thing is this supports an addition to our spec to allow
a bundle author to create a 00_deploy-done.sh script to test whether
their services are up and can be acted upon.

Second item uses a new shell function in conjure-up to print out a
current leader in a group of units.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>